+++ recipe/passwordless/passwordless_email_test.go
@@ -17,6 +17,8 @@
 package passwordless
 
 import (
+	"encoding/json"
+	"io/ioutil"
 	"net/http"
 	"testing"
 
@@ -55,6 +57,24 @@ func TestDefaultBackwardCompatibilityPasswordlessLogin(t *testing.T) {
 	}
 
 	resp, err := unittesting.PasswordlessEmailLoginRequest("test@example.com", testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
+	assert.True(t, PasswordlessLoginEmailSentForTest)
+	assert.Equal(t, PasswordlessLoginEmailDataForTest.Email, "test@example.com")
+	assert.NotNil(t, PasswordlessLoginEmailDataForTest.UrlWithLinkCode)
+	assert.NotNil(t, PasswordlessLoginEmailDataForTest.UserInputCode)
+
+	// Test resend
+	ResetForTest()
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
 	assert.NoError(t, err)
 	assert.Equal(t, resp.StatusCode, http.StatusOK)
 
@@ -107,6 +127,13 @@ func TestBackwardCompatibilityPasswordlessLogin(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
 	// Default handler not called
 	assert.False(t, PasswordlessLoginEmailSentForTest)
 	assert.Empty(t, PasswordlessLoginEmailDataForTest.Email)
@@ -114,6 +141,23 @@ func TestBackwardCompatibilityPasswordlessLogin(t *testing.T) {
 	assert.Nil(t, PasswordlessLoginEmailDataForTest.UrlWithLinkCode)
 
 	// Custom handler called
+	assert.Equal(t, plessEmail, "test@example.com")
+	assert.NotNil(t, code)
+	assert.NotNil(t, urlWithCode)
+	assert.NotZero(t, codeLife)
+	assert.True(t, customCalled)
+
+	// Test resend
+	customCalled = false
+	plessEmail = ""
+	code = nil
+	urlWithCode = nil
+	codeLife = 0
+
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+
 	assert.Equal(t, plessEmail, "test@example.com")
 	assert.NotNil(t, code)
 	assert.NotNil(t, urlWithCode)
@@ -171,6 +215,13 @@ func TestCustomOverridePasswordlessLogin(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
 	// Default handler not called
 	assert.False(t, PasswordlessLoginEmailSentForTest)
 	assert.Empty(t, PasswordlessLoginEmailDataForTest.Email)
@@ -178,6 +229,23 @@ func TestCustomOverridePasswordlessLogin(t *testing.T) {
 	assert.Nil(t, PasswordlessLoginEmailDataForTest.UrlWithLinkCode)
 
 	// Custom handler called
+	assert.Equal(t, plessEmail, "test@example.com")
+	assert.NotNil(t, code)
+	assert.NotNil(t, urlWithCode)
+	assert.NotZero(t, codeLife)
+	assert.True(t, customCalled)
+
+	// Test resend
+	customCalled = false
+	plessEmail = ""
+	code = nil
+	urlWithCode = nil
+	codeLife = 0
+
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+
 	assert.Equal(t, plessEmail, "test@example.com")
 	assert.NotNil(t, code)
 	assert.NotNil(t, urlWithCode)
@@ -254,11 +322,37 @@ func TestSMTPOverridePasswordlessLogin(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
 	// Default handler not called
 	assert.False(t, PasswordlessLoginEmailSentForTest)
 	assert.Empty(t, PasswordlessLoginEmailDataForTest.Email)
 	assert.Nil(t, PasswordlessLoginEmailDataForTest.UserInputCode)
 	assert.Nil(t, PasswordlessLoginEmailDataForTest.UrlWithLinkCode)
+
+	assert.Equal(t, plessEmail, "test@example.com")
+	assert.NotNil(t, code)
+	assert.NotNil(t, urlWithCode)
+	assert.NotZero(t, codeLife)
+	assert.Equal(t, getContentCalled, true)
+	assert.Equal(t, sendRawEmailCalled, true)
+
+	// Test resend
+	getContentCalled = false
+	sendRawEmailCalled = false
+	plessEmail = ""
+	code = nil
+	urlWithCode = nil
+	codeLife = 0
+
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
 	assert.Equal(t, plessEmail, "test@example.com")
 	assert.NotNil(t, code)