+++ recipe/passwordless/passwordless_sms_test.go
@@ -17,6 +17,8 @@
 package passwordless
 
 import (
+	"encoding/json"
+	"io/ioutil"
 	"net/http"
 	"testing"
 
@@ -55,6 +57,24 @@ func TestSmsDefaultBackwardCompatibilityPasswordlessLogin(t *testing.T) {
 	}
 
 	resp, err := unittesting.PasswordlessPhoneLoginRequest("+919876543210", testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
+	assert.True(t, PasswordlessLoginSmsSentForTest)
+	assert.Equal(t, PasswordlessLoginSmsDataForTest.Phone, "+919876543210")
+	assert.NotNil(t, PasswordlessLoginSmsDataForTest.UrlWithLinkCode)
+	assert.NotNil(t, PasswordlessLoginSmsDataForTest.UserInputCode)
+
+	// Test resend
+	ResetForTest()
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
 	assert.NoError(t, err)
 	assert.Equal(t, resp.StatusCode, http.StatusOK)
 
@@ -107,6 +127,13 @@ func TestSmsBackwardCompatibilityPasswordlessLogin(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
 	// Default handler not called
 	assert.False(t, PasswordlessLoginSmsSentForTest)
 	assert.Empty(t, PasswordlessLoginSmsDataForTest.Phone)
@@ -114,6 +141,23 @@ func TestSmsBackwardCompatibilityPasswordlessLogin(t *testing.T) {
 	assert.Nil(t, PasswordlessLoginSmsDataForTest.UrlWithLinkCode)
 
 	// Custom handler called
+	assert.Equal(t, plessPhone, "+919876543210")
+	assert.NotNil(t, code)
+	assert.NotNil(t, urlWithCode)
+	assert.NotZero(t, codeLife)
+	assert.True(t, customCalled)
+
+	// Test resend
+	customCalled = false
+	plessPhone = ""
+	code = nil
+	urlWithCode = nil
+	codeLife = 0
+
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+
 	assert.Equal(t, plessPhone, "+919876543210")
 	assert.NotNil(t, code)
 	assert.NotNil(t, urlWithCode)
@@ -171,6 +215,13 @@ func TestSmsCustomOverridePasswordlessLogin(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
 	// Default handler not called
 	assert.False(t, PasswordlessLoginSmsSentForTest)
 	assert.Empty(t, PasswordlessLoginSmsDataForTest.Phone)
@@ -178,6 +229,23 @@ func TestSmsCustomOverridePasswordlessLogin(t *testing.T) {
 	assert.Nil(t, PasswordlessLoginSmsDataForTest.UrlWithLinkCode)
 
 	// Custom handler called
+	assert.Equal(t, plessPhone, "+919876543210")
+	assert.NotNil(t, code)
+	assert.NotNil(t, urlWithCode)
+	assert.NotZero(t, codeLife)
+	assert.True(t, customCalled)
+
+	// Test resend
+	customCalled = false
+	plessPhone = ""
+	code = nil
+	urlWithCode = nil
+	codeLife = 0
+
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+
 	assert.Equal(t, plessPhone, "+919876543210")
 	assert.NotNil(t, code)
 	assert.NotNil(t, urlWithCode)
@@ -252,11 +320,37 @@ func TestSmsTwilioOverridePasswordlessLogin(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	assert.NoError(t, err)
+	body := map[string]string{}
+
+	err = json.Unmarshal(bodyBytes, &body)
+	assert.NoError(t, err)
+
 	// Default handler not called
 	assert.False(t, PasswordlessLoginSmsSentForTest)
 	assert.Empty(t, PasswordlessLoginSmsDataForTest.Phone)
 	assert.Nil(t, PasswordlessLoginSmsDataForTest.UserInputCode)
 	assert.Nil(t, PasswordlessLoginSmsDataForTest.UrlWithLinkCode)
+
+	assert.Equal(t, plessPhone, "+919876543210")
+	assert.NotNil(t, code)
+	assert.NotNil(t, urlWithCode)
+	assert.NotZero(t, codeLife)
+	assert.Equal(t, getContentCalled, true)
+	assert.Equal(t, sendRawSmsCalled, true)
+
+	// Test resend
+	getContentCalled = false
+	sendRawSmsCalled = false
+	plessPhone = ""
+	code = nil
+	urlWithCode = nil
+	codeLife = 0
+
+	resp, err = unittesting.PasswordlessLoginResendRequest(body["deviceId"], body["preAuthSessionId"], testServer.URL)
+	assert.NoError(t, err)
+	assert.Equal(t, http.StatusOK, resp.StatusCode)
 
 	assert.Equal(t, plessPhone, "+919876543210")
 	assert.NotNil(t, code)